pyproject.toml
[tool.poetry]
name = "probe-design"
version = "0.1.0"
description = ""
authors = ["Quentin Verron "]
maintainers = ["Zafer Kosar"]
license = "n"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
numpy = "^1.26.2"
0 = "^0.0.0"
pandas = "^2.1.4"
tabulate = "^0.9.0"
biopython = "^1.82"
tqdm = "^4.66.1"
joblib = "^1.3.2"
click = "^8.1.7"
poetry = "^1.7.1"

[tool.poetry.group.dev.dependencies]
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
